게시글 비밀번호 기능 추가
비밀번호 기능 추가 및 세션에 보안 관련 코드 추가

+++ src/main/java/com/hansol/board/post/controller/PostController.java
@@ -7,11 +7,15 @@ import com.hansol.board.post.domain.Post;
 import com.hansol.board.post.domain.PostEntity;
 import com.hansol.board.post.form.EditPostForm;
 import com.hansol.board.post.form.SavePostForm;
+import com.hansol.board.post.request.CheckPassword;
 import com.hansol.board.post.response.EditorResponse;
 import com.hansol.board.post.service.PostService;
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -97,14 +101,27 @@ public class PostController {
     @GetMapping("{code}/edit/{id}")
     public String editForm(@PathVariable String code,
                            @PathVariable Long id,
-                           Model model) {
+                           Model model,
+                           HttpSession session,
+                           RedirectAttributes redirectAttributes) {
         Optional<Post> findPost = postService.findPostById(id);
         Post post = findPost.orElseThrow(NoPostException::new);
         EditPostForm form = EditPostForm.from(post);
 
         model.addAttribute("form", form);
         model.addAttribute("boards", boardInfoRepository.findAll());
-        return "/board-skin/" + code + "/edit";
+
+        Boolean auth = (Boolean) session.getAttribute("auth");
+        Long sessionPostId = (Long) session.getAttribute("id");
+        String sessionType = (String) session.getAttribute("type");
+
+        if (auth != null && auth && sessionPostId.equals(id) && sessionType.equals("edit")) {
+            return "/board-skin/" + code + "/edit";
+        } else {
+            redirectAttributes.addAttribute("code", code);
+            redirectAttributes.addAttribute("id", id);
+            return "redirect:/post/{code}/view/{id}";
+        }
     }
 
     @PostMapping("{code}/edit/{id}")
@@ -113,13 +130,21 @@ public class PostController {
                              @PathVariable String code,
                              @PathVariable Long id,
                              Model model,
-                             RedirectAttributes redirectAttributes) {
+                             RedirectAttributes redirectAttributes,
+                             HttpSession session) {
         if (bindingResult.hasErrors()) {
             model.addAttribute("boards", boardInfoRepository.findAll());
             return "/board-skin/" + code + "/edit";
         }
 
-        postService.update(Post.formEditForm(editPostForm), editPostForm.getPassword());
+        Boolean auth = (Boolean) session.getAttribute("auth");
+        Long sessionPostId = (Long) session.getAttribute("id");
+        String sessionType = (String) session.getAttribute("type");
+
+        if (auth != null && auth && sessionPostId.equals(id) && sessionType.equals("edit")) {
+            postService.update(Post.formEditForm(editPostForm));
+        }
+
         redirectAttributes.addAttribute("code", code);
         redirectAttributes.addAttribute("id", id);
         return "redirect:/post/{code}/view/{id}";
@@ -128,14 +153,26 @@ public class PostController {
     @PostMapping("{code}/delete/{id}")
     public String deletePost(@PathVariable String code,
                              @PathVariable Long id,
-                             RedirectAttributes redirectAttributes) {
-        Optional<Post> findPost = postService.findPostById(id);
-        findPost.orElseThrow(NoPostException::new);
-        Post post = findPost.get();
-        postService.remove(id, post.getPassword());
+                             RedirectAttributes redirectAttributes,
+                             HttpSession session) {
 
-        redirectAttributes.addAttribute("code", code);
-        return "redirect:/post/{code}/list";
+        Boolean auth = (Boolean) session.getAttribute("auth");
+        Long sessionPostId = (Long) session.getAttribute("id");
+        String sessionType = (String) session.getAttribute("type");
+
+        if (auth != null && auth && sessionPostId.equals(id) && sessionType.equals("delete")) {
+            Optional<Post> findPost = postService.findPostById(id);
+            findPost.orElseThrow(NoPostException::new);
+            Post post = findPost.get();
+            postService.remove(id, post.getPassword());
+
+            redirectAttributes.addAttribute("code", code);
+            return "redirect:/post/{code}/list";
+        } else {
+            redirectAttributes.addAttribute("code", code);
+            redirectAttributes.addAttribute("id", id);
+            return "redirect:/post/{code}/view/{id}";
+        }
     }
 
     @ResponseBody
@@ -168,5 +205,20 @@ public class PostController {
 
         }
         return null;
+    }
+
+    @ResponseBody
+    @PostMapping("chkPwd")
+    public ResponseEntity<String> checkPassword(@RequestBody CheckPassword request,
+                                                HttpSession session) {
+        Boolean result = postService.checkPassword(request);
+        if (result) {
+            session.setAttribute("auth", true);
+            session.setAttribute("type", request.getType());
+            session.setAttribute("id", request.getId());
+            return ResponseEntity.ok("success");
+        } else {
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("잘못된 비밀번호입니다.");
+        }
     }
 }
